+++ src/index.d.ts
@@ -4,15 +4,32 @@ import { SnackbarContentProps } from '@material-ui/core/SnackbarContent';
 
 type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>
 type ClassNameMap<ClassKey extends string = string> = Record<ClassKey, string>;
+type Modify<T, R> = Pick<T, Exclude<keyof T, keyof R>> & R
 
+type RemovedAttributes = 'open' | 'message' | 'classes';
+export type SnackbarKey = string | number;
 export type VariantType = 'default' | 'error' | 'success' | 'warning' | 'info';
 
 export type SnackbarMessage = string | React.ReactNode;
-export type SnackbarAction = SnackbarContentProps['action'] | ((key: OptionsObject['key']) => React.ReactNode);
-export type SnackbarContent = React.ReactNode | ((key: OptionsObject['key'], message: SnackbarMessage) => React.ReactNode);
+export type SnackbarAction = SnackbarContentProps['action'] | ((key: SnackbarKey) => React.ReactNode);
+export type SnackbarContent = React.ReactNode | ((key: SnackbarKey, message: SnackbarMessage) => React.ReactNode);
 
-export interface OptionsObject extends Omit<SnackbarProps, 'open' | 'message' | 'classes'> {
-    key?: string | number;
+export type CloseReason = 'timeout' | 'clickaway' | 'maxsnack' | 'instructed';
+export type TransitionCloseHandler = (event: React.SyntheticEvent<any> | null, reason: CloseReason, key: SnackbarKey) => void;
+export type TransitionEnterHandler = (node: HTMLElement, isAppearing: boolean, key: SnackbarKey) => void;
+export type TransitionHandler = (node: HTMLElement, key: SnackbarKey) => void;
+export type TransitionHandlerProps = {
+    onClose?: TransitionCloseHandler;
+    onEnter?: TransitionHandler;
+    onEntering?: TransitionHandler;
+    onEntered?: TransitionEnterHandler;
+    onExit?: TransitionHandler;
+    onExiting?: TransitionHandler;
+    onExited?: TransitionHandler;
+}
+
+export interface OptionsObject extends Modify<Omit<SnackbarProps, RemovedAttributes>, TransitionHandlerProps> {
+    key?: SnackbarKey;
     variant?: VariantType;
     persist?: boolean;
     preventDuplicate?: boolean;
@@ -33,8 +50,8 @@ export type VariantClassKey = 'variantSuccess' | 'variantError' | 'variantInfo' 
 export type CombinedClassKey = VariantClassKey | ContainerClassKey | SnackbarClassKey;
 
 export interface WithSnackbarProps {
-    enqueueSnackbar: (message: SnackbarMessage, options?: OptionsObject) => OptionsObject['key'];
-    closeSnackbar: (key?: OptionsObject['key']) => void;
+    enqueueSnackbar: (message: SnackbarMessage, options?: OptionsObject) => SnackbarKey;
+    closeSnackbar: (key?: SnackbarKey) => void;
 }
 
 export function withSnackbar<P extends WithSnackbarProps>(component: React.ComponentType<P>):
@@ -44,7 +61,7 @@ export function withSnackbar<P extends WithSnackbarProps>(component: React.Compo
 export function useSnackbar(): WithSnackbarProps;
 
 // all material-ui props, including class keys for notistack and material-ui with additional notistack props
-export interface SnackbarProviderProps extends Omit<SnackbarProps, 'open' | 'message' | 'classes'> {
+export interface SnackbarProviderProps extends Modify<Omit<SnackbarProps, RemovedAttributes>, TransitionHandlerProps> {
     classes?: Partial<ClassNameMap<CombinedClassKey>>;
     maxSnack?: number;
     iconVariant?: Partial<Record<VariantType, React.ReactNode>>;